+++ src/extension/CCResponseFixable.ts
@@ -9,6 +9,7 @@ import { fixUE4Optimization } from "./fixes/optimizeUE4";
 import { fixWrongCppStandard } from "./fixes/wrongCppStandard";
 
 import * as console from "../console";
+import { fixLaunchFile } from "./fixes/launch";
 
 
 
@@ -38,6 +39,9 @@ export class CCResponseFixable extends Fixable {
 
         fixWrongCppStandard(this.project);
         console.log("End fix wrong cppStandard.\n");
+
+        fixLaunchFile(this.project);
+        console.log("End fix launch.json.\n");
 
         return;
     }